Add Lefthook as a git hook runner
It is useful to ensure that all Typescript files compile without error
and that there are no linting errors on each commit, and to ensure
there are no vulnerabilities in dependencies before pushing.

Moving forward, we can ensure that all tests pass before pushing.

Lefthook is written in Go, so it is faster than "husky", the traditional
git hook tool used in the JS world. It can run tasks in parallel and
works for many different programming languages.

- Ensure commit messages are not terrible
- Run `yarn audit` before push
- Compile typescript and run ESLint before each commit