Add ability to add new addresses & list addresses
Adding an address is different than generating an address. Adding an
address implies you already have the address and secret, and want to
be able to use it through the CLI.

A good example would be an address funded from the Testnet Faucet:
https://xrpl.org/xrp-testnet-faucet.html

Listing addresses is useful to know what addresses the CLI already
has access to, and as proof that we are persisting addresses in local
configuration.